Move build file to mjs and update packages
